Issue #25 Rename connect package to network
