+++ src/main/java/systems/composable/dropwizard/cassandra/network/AddressTranslatorFactory.java
@@ -1,4 +1,4 @@
-package systems.composable.dropwizard.cassandra.connect;
+package systems.composable.dropwizard.cassandra.network;
 
 import com.datastax.driver.core.policies.AddressTranslator;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
@@ -9,7 +9,7 @@ import io.dropwizard.jackson.Discoverable;
  * <p>
  * <p>To create your own, just:
  * <p>
- * <ol> <li>Create a class which implements {@link systems.composable.dropwizard.cassandra.connect.AddressTranslatorFactory}. <li>Annotate it with
+ * <ol> <li>Create a class which implements {@link systems.composable.dropwizard.cassandra.network.AddressTranslatorFactory}. <li>Annotate it with
  * {@code @JsonTypeName} and give it a unique type name. <li>Add a {@code META-INF/services/systems.composable.dropwizard.cassandra.connect.AddressTranslatorFactory}
  * file with your implementation's full class name to the class path. </ol>
  *
