+++ src/main/resources/META-INF/services/io.dropwizard.jackson.Discoverable
@@ -4,4 +4,4 @@ systems.composable.dropwizard.cassandra.retry.RetryPolicyFactory
 systems.composable.dropwizard.cassandra.loadbalancing.LoadBalancingPolicyFactory
 systems.composable.dropwizard.cassandra.speculativeexecution.SpeculativeExecutionPolicyFactory
 systems.composable.dropwizard.cassandra.ssl.SSLOptionsFactory
-systems.composable.dropwizard.cassandra.connect.AddressTranslatorFactory
+network.AddressTranslatorFactory
