+++ src/main/resources/META-INF/services/systems.composable.dropwizard.cassandra.connect.AddressTranslatorFactory
@@ -1,2 +1,2 @@
-systems.composable.dropwizard.cassandra.connect.EC2MultiRegionAddressTranslatorFactory
-systems.composable.dropwizard.cassandra.connect.IdentityTranslatorFactory
+systems.composable.dropwizard.cassandra.network.EC2MultiRegionAddressTranslatorFactory
+systems.composable.dropwizard.cassandra.network.IdentityTranslatorFactory
