+++ src/test/resources/smoke/network/addressTranslator-ec2.yml
