+++ src/test/resources/smoke/network/addressTranslator-identity.yml
